+++ peachjam/views/documents.py
@@ -1,5 +1,4 @@
 from cobalt import FrbrUri
-from django.conf import settings
 from django.http import Http404, HttpResponse
 from django.shortcuts import get_object_or_404, redirect, reverse
 from django.utils.decorators import method_decorator
@@ -9,111 +8,7 @@ from django.views.generic import DetailView, View
 from peachjam.helpers import add_slash, add_slash_to_frbr_uri
 from peachjam.models import CoreDocument
 from peachjam.registry import registry
-
-
-class RedirectResolver:
-    RESOLVER_MAPPINGS = {
-        "africanlii": {
-            "country_code": "aa",
-            "domain": "africanlii.org",
-        },
-        "eswatinilii": {
-            "country_code": "sz",
-            "domain": "eswatinilii.org",
-        },
-        "ghalii": {
-            "country_code": "gh",
-            "domain": "ghalii.org",
-        },
-        "lawlibrary": {
-            "country_code": "za",
-            "domain": "lawlibrary.org.za",
-        },
-        "leslii": {
-            "country_code": "ls",
-            "domain": "lesotholii.org",
-        },
-        "malawilii": {
-            "country_code": "mw",
-            "domain": "malawilii.org",
-        },
-        "mauritiuslii": {
-            "country_code": "mu",
-            "domain": "mauritiuslii.org",
-        },
-        "namiblii": {
-            "country_code": "na",
-            "domain": "namiblii.org",
-        },
-        "nigerialii": {
-            "country_code": "ng",
-            "domain": "nigerialii.org",
-        },
-        "open by-laws": {
-            "place_code": [],
-            "domain": "openbylaws.org.za",
-        },
-        "rwandalii": {
-            "country_code": "rw",
-            "domain": "rwandalii.org",
-        },
-        "seylii": {
-            "country_code": "sc",
-            "domain": "seylii.org",
-        },
-        "sierralii": {
-            "country_code": "sl",
-            "domain": "sierralii.org",
-        },
-        "tanzlii": {
-            "country_code": "tz",
-            "domain": "tanzlii.org",
-        },
-        "tcilii": {
-            "country_code": "tc",
-            "domain": "tcilii.org",
-        },
-        "ulii": {
-            "country_code": "ug",
-            "domain": "ulii.org",
-        },
-        "zambialii": {
-            "country_code": "zm",
-            "domain": "zambialii.org",
-        },
-        "zanzibarlii": {
-            "place_code": "tz-znz",
-            "domain": "zanzibarlii.org",
-        },
-        "zimlii": {
-            "country_code": "zw",
-            "domain": "zimlii.org",
-        },
-    }
-
-    def __init__(self, app_name):
-        self.current_authority = self.RESOLVER_MAPPINGS[app_name.lower()]
-
-    def get_domain_for_frbr_uri(self, parsed_uri):
-        best_domain = self.get_best_domain(parsed_uri)
-        if best_domain != self.current_authority["domain"]:
-            return best_domain
-        return None
-
-    def get_best_domain(self, parsed_uri):
-        country_code = parsed_uri.country
-        place_code = parsed_uri.place
-
-        if country_code != place_code:
-            for key, mapping in self.RESOLVER_MAPPINGS.items():
-                if mapping.get("place_code") == place_code:
-                    return mapping.get("domain")
-
-        # if no domain matching with place code is found use country code
-        for key, mapping in self.RESOLVER_MAPPINGS.items():
-            if mapping.get("country_code") == country_code:
-                return mapping.get("domain")
-        return None
+from peachjam.resolver import resolver
 
 
 class DocumentDetailViewResolver(View):
@@ -145,10 +40,8 @@ class DocumentDetailViewResolver(View):
         )
 
         if not obj:
-            resolver = RedirectResolver(settings.PEACHJAM["APP_NAME"])
-            domain = resolver.get_domain_for_frbr_uri(parsed_frbr_uri)
-            if domain:
-                url = f"https://{domain}{frbr_uri}"
+            url = resolver.get_url_for_frbr_uri(parsed_frbr_uri, frbr_uri)
+            if url:
                 return redirect(url)
             raise Http404()
