+++ pyproject.toml
@@ -32,6 +32,7 @@ dependencies = [
     "django-background-tasks>=1.2.5",
     "django-ckeditor>=6.4.2",
     "django-compressor>=3.1",
+    "django-cors-headers>=4.3.1",
     "django-countries-plus>=1.3.2",
     "django-debug-toolbar>=3.2.4,<4.2.0",
     "django-elasticsearch-debug-toolbar>=3.0.2",
